requirements.txt
# Core dependencies
sqlalchemy>=2.0.0
asyncpg>=0.27.0
aiosqlite>=0.17.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
loguru>=0.6.0
python-dotenv>=1.0.0
alembic>=1.9.0
httpx>=0.24.0
fastapi>=0.95.0
uvicorn>=0.20.0
pandas>=1.5.0
numpy>=1.22.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0

# Development
black>=23.1.0
isort>=5.12.0
mypy>=1.0.0
flake8>=6.0.0